+++ src/uplinksinrtesterimpl.h
@@ -55,7 +55,7 @@ typedef std::map<ChannelRNTI, bool> RNTIChannelSINRResults;
 class UplinkSINRTesterImpl : public SINRTesterImpl
 {
 public:
-  UplinkSINRTesterImpl(double sinr, double noiseFloor, std::uint64_t carrierFrequencyHz) : 
+  UplinkSINRTesterImpl(float sinr, float noiseFloor, std::uint64_t carrierFrequencyHz) : 
    sinr_dB_{sinr},
    noiseFloor_dBm_{noiseFloor},
    carrierFrequencyHz_{carrierFrequencyHz}
@@ -70,8 +70,8 @@ public:
   SINRTester::SINRTesterResult sinrCheck(CHANNEL_TYPE ctype, uint16_t rnti) override;  
 
 private:
-  const double sinr_dB_;
-  const double noiseFloor_dBm_;
+  const float sinr_dB_;
+  const float noiseFloor_dBm_;
   std::uint64_t carrierFrequencyHz_;
 };
